+++ src/Views/MainWindow/MainWindow.ContextMenu.cs
@@ -1,12 +1,27 @@
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
+using System.Windows.Interop;
 using System.Windows.Media;
 
 namespace Tfx;
 
 public partial class MainWindow
 {
+    private void OpenWithDialog(string path)
+    {
+        try
+        {
+            var hwnd = new WindowInteropHelper(this).Handle;
+            ShellOpenWith.Show(hwnd, path);
+        }
+        catch (Exception ex)
+        {
+            SetStatus(Loc.F("Open with failed: {0}", ex.Message));
+        }
+    }
+
+
     private void Grid_PreviewMouseRightButtonDown(object sender, MouseButtonEventArgs e)
     {
         if (sender is not DataGrid grid)
@@ -50,8 +65,11 @@ public partial class MainWindow
         var selection = ActiveSelectedItems().Where(i => !i.IsParent).ToArray();
         var hasSelection = selection.Length > 0;
         var oneSelected = selection.Length == 1;
-        var hasZipSelection = selection.Any(i => !i.IsDirectory && System.IO.Path.GetExtension(i.FullPath).Equals(".zip", StringComparison.OrdinalIgnoreCase));
+        var hasZipSelection = selection.Any(i => !i.IsDirectory && System.IO.Path.GetExtension(i.FullPath).Equals(".zip", StringComparison.OrdinalIgnoreCase) && !ArchivePath.Contains(i.FullPath));
         var hasClipboard = Clipboard.ContainsFileDropList();
+        var inArchive = ArchivePath.Contains(GetCurrentPath(grid));
+        var selectionHasArchive = selection.Any(s => ArchivePath.Contains(s.FullPath));
+        var writableContext = !inArchive && !selectionHasArchive;
 
         var open = new MenuItem { Header = Loc.T("Open"), InputGestureText = "Enter", IsEnabled = oneSelected };
         open.Click += (_, _) =>
@@ -63,11 +81,22 @@ public partial class MainWindow
         };
         menu.Items.Add(open);
 
+        var openWithEnabled = oneSelected && !selection[0].IsDirectory;
+        var openWith = new MenuItem { Header = Loc.T("Open with..."), IsEnabled = openWithEnabled };
+        openWith.Click += (_, _) =>
+        {
+            if (selection.Length == 1 && !selection[0].IsDirectory)
+            {
+                OpenWithDialog(selection[0].FullPath);
+            }
+        };
+        menu.Items.Add(openWith);
+
         var reveal = new MenuItem { Header = Loc.T("Reveal in Explorer") };
         reveal.Click += (_, _) => RevealInExplorer();
         menu.Items.Add(reveal);
 
-        var pinTargetIsDir = oneSelected && selection[0].IsDirectory;
+        var pinTargetIsDir = oneSelected && selection[0].IsDirectory && !ArchivePath.Contains(selection[0].FullPath);
         var pinTargetPath = pinTargetIsDir ? selection[0].FullPath : null;
         var pinAlreadyPinned = pinTargetPath != null && _pinned.Contains(pinTargetPath);
         var pin = new MenuItem
@@ -95,7 +124,7 @@ public partial class MainWindow
 
         menu.Items.Add(new Separator());
 
-        var cut = new MenuItem { Header = Loc.T("Cut"), InputGestureText = "Ctrl+X", IsEnabled = hasSelection };
+        var cut = new MenuItem { Header = Loc.T("Cut"), InputGestureText = "Ctrl+X", IsEnabled = hasSelection && writableContext };
         cut.Click += (_, _) => CopySelection(true);
         menu.Items.Add(cut);
 
@@ -103,21 +132,41 @@ public partial class MainWindow
         copy.Click += (_, _) => CopySelection(false);
         menu.Items.Add(copy);
 
-        var paste = new MenuItem { Header = Loc.T("Paste"), InputGestureText = "Ctrl+V", IsEnabled = hasClipboard };
+        var paste = new MenuItem { Header = Loc.T("Paste"), InputGestureText = "Ctrl+V", IsEnabled = hasClipboard && !inArchive };
         paste.Click += (_, _) => PasteIntoActivePane();
         menu.Items.Add(paste);
 
-        var compress = new MenuItem { Header = Loc.T("Compress to Zip"), InputGestureText = "Ctrl+K", IsEnabled = hasSelection };
+        var copyCurrentPath = new MenuItem { Header = Loc.T("Copy current path"), IsEnabled = oneSelected };
+        copyCurrentPath.Click += (_, _) => CopySelectedPath(selection);
+        menu.Items.Add(copyCurrentPath);
+
+        menu.Items.Add(new Separator());
+
+        var compress = new MenuItem { Header = Loc.T("Compress to Zip"), InputGestureText = "Ctrl+K", IsEnabled = hasSelection && writableContext };
         compress.Click += (_, _) => CompressSelection();
         menu.Items.Add(compress);
 
-        var extract = new MenuItem { Header = Loc.T("Extract Zip"), InputGestureText = "Ctrl+Shift+E", IsEnabled = hasZipSelection };
+        var extract = new MenuItem { Header = Loc.T("Extract Zip"), InputGestureText = "Ctrl+Shift+E", IsEnabled = hasZipSelection && writableContext };
         extract.Click += (_, _) => ExtractSelectedArchives();
         menu.Items.Add(extract);
 
         menu.Items.Add(new Separator());
 
-        var rename = new MenuItem { Header = Loc.T("Rename"), InputGestureText = "F2", IsEnabled = oneSelected };
+        var newFolder = new MenuItem { Header = Loc.T("New Folder"), InputGestureText = "Ctrl+N", IsEnabled = !inArchive };
+        newFolder.Click += (_, _) => NewFolder();
+        menu.Items.Add(newFolder);
+
+        var newFile = new MenuItem { Header = Loc.T("New File"), InputGestureText = "Ctrl+Shift+N", IsEnabled = !inArchive };
+        newFile.Click += (_, _) => NewFile();
+        menu.Items.Add(newFile);
+
+        var openTerminal = new MenuItem { Header = Loc.T("Open Terminal here"), InputGestureText = "Ctrl+Shift+T", IsEnabled = !inArchive };
+        openTerminal.Click += (_, _) => OpenTerminal();
+        menu.Items.Add(openTerminal);
+
+        menu.Items.Add(new Separator());
+
+        var rename = new MenuItem { Header = Loc.T("Rename"), InputGestureText = "F2", IsEnabled = oneSelected && writableContext };
         rename.Click += (_, _) =>
         {
             if (grid.SelectedItem is FileItem item && !item.IsParent)
@@ -127,31 +176,13 @@ public partial class MainWindow
         };
         menu.Items.Add(rename);
 
-        var trash = new MenuItem { Header = Loc.T("Move to Recycle Bin"), InputGestureText = "Del", IsEnabled = hasSelection };
+        var trash = new MenuItem { Header = Loc.T("Move to Recycle Bin"), InputGestureText = "Del", IsEnabled = hasSelection && writableContext };
         trash.Click += (_, _) => MoveSelectionToTrash();
         menu.Items.Add(trash);
 
-        var perm = new MenuItem { Header = Loc.T("Delete permanently"), InputGestureText = "Shift+Del", IsEnabled = hasSelection };
+        var perm = new MenuItem { Header = Loc.T("Delete permanently"), InputGestureText = "Shift+Del", IsEnabled = hasSelection && writableContext };
         perm.Click += (_, _) => DeletePermanently();
         menu.Items.Add(perm);
-
-        menu.Items.Add(new Separator());
-
-        var newFolder = new MenuItem { Header = Loc.T("New Folder"), InputGestureText = "Ctrl+N" };
-        newFolder.Click += (_, _) => NewFolder();
-        menu.Items.Add(newFolder);
-
-        var newFile = new MenuItem { Header = Loc.T("New File"), InputGestureText = "Ctrl+Shift+N" };
-        newFile.Click += (_, _) => NewFile();
-        menu.Items.Add(newFile);
-
-        var openTerminal = new MenuItem { Header = Loc.T("Open Terminal here"), InputGestureText = "Ctrl+Shift+T" };
-        openTerminal.Click += (_, _) => OpenTerminal();
-        menu.Items.Add(openTerminal);
-
-        var copyCurrentPath = new MenuItem { Header = Loc.T("Copy current path"), IsEnabled = oneSelected };
-        copyCurrentPath.Click += (_, _) => CopySelectedPath(selection);
-        menu.Items.Add(copyCurrentPath);
 
         return menu;
     }